weatherlookbook/accountapp/views.py
from django.http.response import JsonResponse
from django.shortcuts import render
from rest_framework.decorators import api_view, permission_classes, authentication_classes
from rest_framework.permissions import IsAuthenticated
from rest_framework_jwt.authentication import JSONWebTokenAuthentication
# Create your views here.

@api_view(['POST'])
@permission_classes((IsAuthenticated, ))
@authentication_classes((JSONWebTokenAuthentication,))
def join(request):
    context = {'msg':'good'}
    return JsonResponse(context)

def login(request):
    context = {'msg':'good'}
    return JsonResponse(context)
    